Dynamic Workload Generator and add Applications (#93)
* Add Collector

* Add application Bolt(TP,OB) finish dynamic scheduler

* Finish dynamic scheduler

* Only the thread-0

* Fix concurrent access bug

* Add a configuration

* System sensitivity study (#86)

* unified the namespace, make it be consistent with terms in paper. todo: update scripts.

* update

* add scripts for system sensitivity study

* code clean

* updated countdown to atomic integer to fix the bug for DFS

* refactor data structures, operations are extending the same operation now.

* update\

* unified the namespace, make it be consistent with terms in paper. todo: update scripts.

* update

* add scripts for system sensitivity study

* code clean

* updated countdown to atomic integer to fix the bug for DFS

* refactor data structures, operations are extending the same operation now.

* update\

* code clean

* update experiment scripts

* unified the namespace, make it be consistent with terms in paper. todo: update scripts.

* update

* add scripts for system sensitivity study

* updated countdown to atomic integer to fix the bug for DFS

* refactor data structures, operations are extending the same operation now.

* update\

* code clean

* unified the namespace, make it be consistent with terms in paper. todo: update scripts.

* update

* add scripts for system sensitivity study

* updated countdown to atomic integer to fix the bug for DFS

* refactor data structures, operations are extending the same operation now.

* update\

* update experiment scripts

* solve conflict


* solve conflict

* Solve the conflict

* Add DynamicWorkloadGenerator

* Add Application(TP and OB)

* Fix some typo

+++ application/src/main/java/common/datatype/PositionReport.java
@@ -87,7 +87,7 @@ public final class PositionReport extends AbstractInputTuple implements IPositio
      * @param position  the horizontal position on the expressway
      */
     public PositionReport(Short time, Integer vid, Integer speed, Integer xway, Short lane, Short direction,
-                          Short segment, Integer position) {
+                          Integer segment, Integer position) {
         super(AbstractLRBTuple.POSITION_REPORT, time, vid);
         assert (speed != null);
         assert (xway != null);
@@ -165,8 +165,8 @@ public final class PositionReport extends AbstractInputTuple implements IPositio
      * @return the VID of this position report
      */
     @Override
-    public final Short getSegment() {
-        return (Short) super.get(SEG_IDX);
+    public final int getSegment() {
+        return (int) super.get(SEG_IDX);
     }
 
     /**
