+++ application/src/main/java/common/datatype/internal/AvgSpeedTuple.java
@@ -124,8 +124,8 @@ public final class AvgSpeedTuple extends StreamValues implements ISegmentIdentif
      * @return the VID of this tuple
      */
     @Override
-    public final Short getSegment() {
-        return (Short) super.get(SEG_IDX);
+    public final int getSegment() {
+        return (Integer) super.get(SEG_IDX);
     }
 
     /**
